requirements.txt
appdirs==1.4.3
click==6.7
docopt==0.6.2
docutils==0.13.1
enum-compat==0.0.2
eventlet==0.20.1
Flask==0.12.1
Flask-MySQLdb==0.2.0
gunicorn==19.7.1
import-string==0.1.0
itsdangerous==0.24
Jinja2==2.9.5
manage==0.1.13
MarkupSafe==1.0
mysql==0.0.1
MySQL-python==1.2.5
mysqlclient==1.3.10
python-dateutil==2.6.0
python-engineio==1.3.1
python-socketio==1.7.4
psycopg2==2.7.1
requests==2.9.1
six==1.10.0
termcolor==1.1.0
Werkzeug==0.12.1
